corrections

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -5,11 +5,11 @@
  */
 int main(void)
 {
-	int limiter = 2;
-	int a = 1;
-	int b = 2;
+	long int limiter = 2;
+	long int a = 1;
+	long int b = 2;
 	long int sum = a + b;
-	int c;
+	long int c;
 
 	printf("1, 2\n");
 	printf("%ld, ", sum);
